+++ datapipe/dsl.py
@@ -3,9 +3,8 @@ from typing import Callable, Dict, List, Tuple, Union
 from abc import ABC
 from dataclasses import dataclass
 
-from datapipe.datatable import DataTable
+from datapipe.datatable import DataStore, DataTable
 from datapipe.store.table_store import TableStore
-from datapipe.metastore import MetaStore
 
 
 @dataclass
@@ -27,21 +26,11 @@ class Catalog:
         self.catalog = catalog
         self.data_tables: Dict[str, DataTable] = {}
 
-    def get_datatable(self, ms: MetaStore, name: str) -> DataTable:
-        if name not in self.data_tables:
-            store = self.catalog[name].store
-            primary_schema = store.get_primary_schema()
-
-            if not primary_schema:
-                raise ValueError("DataFrame must have at least one primary key")
-
-            self.data_tables[name] = DataTable(
-                name=name,
-                meta_table=ms.create_meta_table(name, primary_schema),
-                table_store=store
-            )
-
-        return self.data_tables[name]
+    def get_datatable(self, ds: DataStore, name: str) -> DataTable:
+        return ds.get_or_create_table(
+            name=name,
+            table_store=self.catalog[name].store
+        )
 
 
 class PipelineStep(ABC):
@@ -68,7 +57,7 @@ class BatchGenerate(PipelineStep):
 
 
 @dataclass
-class LabelStudioModeration:
+class LabelStudioModeration(PipelineStep):
     ls_url: str
     inputs: List[str]
     outputs: List[str]