+++ tests/test_datatable.py
@@ -1,5 +1,4 @@
 from typing import cast
-
 import pytest
 
 import cloudpickle
@@ -8,9 +7,8 @@ from sqlalchemy import Column
 from sqlalchemy.sql.sqltypes import Integer
 
 from datapipe.store.database import TableStoreDB
-from datapipe.datatable import DataTable, gen_process, gen_process_many, inc_process, inc_process_many
-from datapipe.metastore import MetaStore
-from datapipe.types import DataDF
+from datapipe.datatable import DataStore, gen_process, gen_process_many, inc_process, inc_process_many
+from datapipe.types import IndexDF, data_to_index
 
 from .util import assert_df_equal, assert_datatable_equal
 
@@ -41,89 +39,87 @@ def yield_df(data):
 
 
 def test_cloudpickle(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(meta_dbconn=dbconn)
 
-    tbl = DataTable(
-        'test',
-        meta_table=ms.create_meta_table('test'),
+    tbl = ds.create_table(
+        name='test',
         table_store=TableStoreDB(dbconn, 'test_data', TEST_SCHEMA, True)
     )
 
-    cloudpickle.dumps([ms, tbl])
+    cloudpickle.dumps([ds, tbl])
 
 
 def test_simple(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'test',
-        meta_table=ms.create_meta_table('test'),
         table_store=TableStoreDB(dbconn, 'test_data', TEST_SCHEMA, True)
     )
 
-    tbl.store(cast(DataDF, TEST_DF))
+    tbl.store_chunk(TEST_DF)
 
     assert_datatable_equal(tbl, TEST_DF)
 
 
 def test_store_less_values(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'test',
-        meta_table=ms.create_meta_table('test'),
-        table_store=TableStoreDB(dbconn, 'test_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'test_data', TEST_SCHEMA, True)
+    )
 
-    tbl.store(TEST_DF)
+    tbl.store_chunk(TEST_DF)
     assert_datatable_equal(tbl, TEST_DF)
 
-    tbl.store(TEST_DF[:5])
+    tbl.store_chunk(TEST_DF[:5], processed_idx=data_to_index(TEST_DF, tbl.primary_keys))
     assert_datatable_equal(tbl, TEST_DF[:5])
 
 
 def test_get_process_ids(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
-    tbl1.store(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
 
-    count, idx_dfs = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_dfs = ds.get_process_ids([tbl1], [tbl2])
     idx = pd.concat(list(idx_dfs))
 
     assert(sorted(list(idx.index)) == list(TEST_DF.index))
 
-    tbl2.store(tbl1.get_data())
+    tbl2.store_chunk(tbl1.get_data())
 
     upd_df = TEST_DF[:5].copy()
     upd_df['a'] += 1
 
     tbl1.store_chunk(upd_df)
 
-    count, idx_dfs = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_dfs = ds.get_process_ids([tbl1], [tbl2])
     idx = pd.concat(list(idx_dfs))
 
     assert_df_equal(idx, upd_df[['id']])
 
 
 def test_gen_process(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1_gen = DataTable(
+    tbl1_gen = ds.create_table(
         'tbl1_gen',
-        meta_table=ms.create_meta_table('tbl1_gen'),
-        table_store=TableStoreDB(dbconn, 'tbl1_gen_data', TEST_SCHEMA, True))
-    tbl1 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_gen_data', TEST_SCHEMA, True)
+    )
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
 
     def gen():
         yield TEST_DF
@@ -165,130 +161,130 @@ def test_gen_process(dbconn) -> None:
 
 
 def test_inc_process_modify_values(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
     def id_func(df):
         return df
 
-    tbl1.store(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
     assert_datatable_equal(tbl2, TEST_DF)
 
     ##########################
-    tbl1.store(TEST_DF_INC1)
+    tbl1.store_chunk(TEST_DF_INC1)
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
     assert_datatable_equal(tbl2, TEST_DF_INC1)
 
 
 def test_inc_process_delete_values_from_input(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
     def id_func(df):
         return df
 
-    tbl1.store(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
     assert_datatable_equal(tbl2, TEST_DF)
 
     ##########################
-    tbl1.store(TEST_DF[:5])
+    tbl1.store_chunk(TEST_DF[:5], processed_idx=data_to_index(TEST_DF, tbl1.primary_keys))
 
-    inc_process(ms, [tbl1], tbl2, id_func, chunksize=2)
+    inc_process(ds, [tbl1], tbl2, id_func, chunksize=2)
 
     assert_datatable_equal(tbl2, TEST_DF[:5])
 
 
 def test_inc_process_delete_values_from_proc(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
     def id_func(df):
         return df[:5]
 
-    tbl2.store(TEST_DF)
+    tbl2.store_chunk(TEST_DF)
 
-    tbl1.store(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
     assert_datatable_equal(tbl2, TEST_DF[:5])
 
 
 def test_inc_process_proc_no_change(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
     def id_func(df):
         return TEST_DF
 
-    tbl2.store(TEST_DF)
-    tbl1.store(TEST_DF)
+    tbl2.store_chunk(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
 
-    count, idx_gen = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_gen = ds.get_process_ids([tbl1], [tbl2])
     idx_dfs = list(idx_gen)
     idx_len = len(pd.concat(idx_dfs)) if len(idx_dfs) > 0 else 0
 
     assert(idx_len == len(TEST_DF))
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
-    count, idx_gen = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_gen = ds.get_process_ids([tbl1], [tbl2])
     idx_dfs = list(idx_gen)
     idx_len = len(pd.concat(idx_dfs)) if len(idx_dfs) > 0 else 0
 
     assert(idx_len == 0)
 
-    tbl1.store(TEST_DF_INC1)
+    tbl1.store_chunk(TEST_DF_INC1)
 
-    count, idx_gen = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_gen = ds.get_process_ids([tbl1], [tbl2])
     idx_dfs = list(idx_gen)
     idx_len = len(pd.concat(idx_dfs)) if len(idx_dfs) > 0 else 0
 
     assert(idx_len == len(TEST_DF))
 
-    inc_process(ms, [tbl1], tbl2, id_func)
+    inc_process(ds, [tbl1], tbl2, id_func)
 
-    count, idx_gen = ms.get_process_ids([tbl1.meta_table], [tbl2.meta_table])
+    count, idx_gen = ds.get_process_ids([tbl1], [tbl2])
     idx_dfs = list(idx_gen)
     idx_len = len(pd.concat(idx_dfs)) if len(idx_dfs) > 0 else 0
 
@@ -299,40 +295,40 @@ def test_inc_process_proc_no_change(dbconn) -> None:
 
 
 def test_gen_process_many(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl_gen = DataTable(
+    tbl_gen = ds.create_table(
         'tbl_gen',
-        meta_table=ms.create_meta_table('tbl_gen'),
-        table_store=TableStoreDB(dbconn, 'tbl_gen_data', TEST_SCHEMA, True))
-    tbl1_gen = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl_gen_data', TEST_SCHEMA, True)
+    )
+    tbl1_gen = ds.create_table(
         'tbl1_gen',
-        meta_table=ms.create_meta_table('tbl1_gen'),
-        table_store=TableStoreDB(dbconn, 'tbl1_gen_data', TEST_SCHEMA, True))
-    tbl2_gen = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_gen_data', TEST_SCHEMA, True)
+    )
+    tbl2_gen = ds.create_table(
         'tbl2_gen',
-        meta_table=ms.create_meta_table('tbl2_gen'),
-        table_store=TableStoreDB(dbconn, 'tbl2_gen_data', TEST_SCHEMA, True))
-    tbl3_gen = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl2_gen_data', TEST_SCHEMA, True)
+    )
+    tbl3_gen = ds.create_table(
         'tbl3_gen',
-        meta_table=ms.create_meta_table('tbl3_gen'),
-        table_store=TableStoreDB(dbconn, 'tbl3_gen_data', TEST_SCHEMA, True))
-    tbl = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl3_gen_data', TEST_SCHEMA, True)
+    )
+    tbl = ds.create_table(
         'tbl',
-        meta_table=ms.create_meta_table('tbl'),
-        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True))
-    tbl1 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True)
+    )
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
-    tbl3 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
+    tbl3 = ds.create_table(
         'tbl3',
-        meta_table=ms.create_meta_table('tbl3'),
-        table_store=TableStoreDB(dbconn, 'tbl3_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl3_data', TEST_SCHEMA, True)
+    )
 
     def gen():
         yield (TEST_DF, TEST_DF_INC1, TEST_DF_INC2, TEST_DF_INC3)
@@ -358,24 +354,24 @@ def test_gen_process_many(dbconn) -> None:
 
 
 def test_inc_process_many_modify_values(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'tbl',
-        meta_table=ms.create_meta_table('tbl'),
-        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True))
-    tbl1 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True)
+    )
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
-    tbl3 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
+    tbl3 = ds.create_table(
         'tbl3',
-        meta_table=ms.create_meta_table('tbl3'),
-        table_store=TableStoreDB(dbconn, 'tbl3_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl3_data', TEST_SCHEMA, True)
+    )
 
     def inc_func(df):
         df1 = df.copy()
@@ -386,16 +382,16 @@ def test_inc_process_many_modify_values(dbconn) -> None:
         df3['a'] += 3
         return df1, df2, df3
 
-    tbl.store(TEST_DF)
+    tbl.store_chunk(TEST_DF)
 
-    inc_process_many(ms, [tbl], [tbl1, tbl2, tbl3], inc_func)
+    inc_process_many(ds, [tbl], [tbl1, tbl2, tbl3], inc_func)
 
     assert_datatable_equal(tbl1, TEST_DF_INC1)
     assert_datatable_equal(tbl2, TEST_DF_INC2)
     assert_datatable_equal(tbl3, TEST_DF_INC3)
 
     ##########################
-    tbl.store(TEST_DF[:5])
+    tbl.store_chunk(TEST_DF[:5], processed_idx=data_to_index(TEST_DF, tbl.primary_keys))
 
     def inc_func_inv(df):
         df1 = df.copy()
@@ -406,7 +402,7 @@ def test_inc_process_many_modify_values(dbconn) -> None:
         df3['a'] += 3
         return df3, df2, df1
 
-    inc_process_many(ms, [tbl], [tbl3, tbl2, tbl1], inc_func_inv)
+    inc_process_many(ds, [tbl], [tbl3, tbl2, tbl1], inc_func_inv)
 
     assert_datatable_equal(tbl1, TEST_DF_INC1[:5])
     assert_datatable_equal(tbl2, TEST_DF_INC2[:5])
@@ -416,18 +412,17 @@ def test_inc_process_many_modify_values(dbconn) -> None:
 
     tbl.store_chunk(TEST_DF[5:])
 
-    inc_process_many(ms, [tbl], [tbl1, tbl2, tbl3], inc_func)
+    inc_process_many(ds, [tbl], [tbl1, tbl2, tbl3], inc_func)
     assert_datatable_equal(tbl1, TEST_DF_INC1)
     assert_datatable_equal(tbl2, TEST_DF_INC2)
     assert_datatable_equal(tbl3, TEST_DF_INC3)
 
 
 def test_inc_process_many_several_inputs(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'tbl',
-        meta_table=ms.create_meta_table('tbl'),
         table_store=TableStoreDB(
             dbconn,
             'tbl_data',
@@ -439,14 +434,14 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
             True
         )
     )
-    tbl1 = DataTable(
+    tbl1 = ds.create_table(
         'tbl1',
-        meta_table=ms.create_meta_table('tbl1'),
-        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True))
-    tbl2 = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
+    )
+    tbl2 = ds.create_table(
         'tbl2',
-        meta_table=ms.create_meta_table('tbl2'),
-        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl2_data', TEST_SCHEMA, True)
+    )
 
     def inc_func(df1, df2):
         df = pd.merge(
@@ -459,10 +454,10 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
         df['a_second'] += 2
         return df
 
-    tbl1.store(TEST_DF)
-    tbl2.store(TEST_DF)
+    tbl1.store_chunk(TEST_DF)
+    tbl2.store_chunk(TEST_DF)
 
-    inc_process_many(ms, [tbl1, tbl2], [tbl], inc_func)
+    inc_process_many(ds, [tbl1, tbl2], [tbl], inc_func)
     assert_datatable_equal(
         tbl,
         pd.DataFrame(
@@ -475,9 +470,9 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
     )
 
     changed_ids = [0, 4, 6]
-    changed_ids_df = pd.DataFrame({'id': changed_ids})
+    changed_ids_df = cast(IndexDF, pd.DataFrame({'id': changed_ids}))
     not_changed_ids = [1, 2, 3, 5, 7, 8, 9]
-    not_changed_ids_df = pd.DataFrame({'id': not_changed_ids})
+    not_changed_ids_df = cast(IndexDF, pd.DataFrame({'id': not_changed_ids}))
 
     tbl2.store_chunk(
         pd.DataFrame(
@@ -488,7 +483,7 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
         )
     )
 
-    inc_process_many(ms, [tbl1, tbl2], [tbl], inc_func)
+    inc_process_many(ds, [tbl1, tbl2], [tbl], inc_func)
 
     assert_df_equal(
         tbl.get_data(idx=changed_ids_df),
@@ -521,7 +516,7 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
         )
     )
 
-    inc_process_many(ms, [tbl1, tbl2], [tbl], inc_func)
+    inc_process_many(ds, [tbl1, tbl2], [tbl], inc_func)
 
     assert_df_equal(
         tbl.get_data(idx=changed_ids_df),
@@ -547,38 +542,38 @@ def test_inc_process_many_several_inputs(dbconn) -> None:
 
 
 def test_inc_process_many_several_outputs(dbconn) -> None:
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
     bad_ids = [0, 1, 5, 8]
     good_ids = [2, 3, 4, 6, 7, 9]
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'tbl',
-        meta_table=ms.create_meta_table('tbl'),
-        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True))
-    tbl_good = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl_data', TEST_SCHEMA, True)
+    )
+    tbl_good = ds.create_table(
         'tbl_good',
-        meta_table=ms.create_meta_table('tbl_good'),
-        table_store=TableStoreDB(dbconn, 'tbl_good_data', TEST_SCHEMA, True))
-    tbl_bad = DataTable(
+        table_store=TableStoreDB(dbconn, 'tbl_good_data', TEST_SCHEMA, True)
+    )
+    tbl_bad = ds.create_table(
         'tbl_bad',
-        meta_table=ms.create_meta_table('tbl_bad'),
-        table_store=TableStoreDB(dbconn, 'tbl_bad_data', TEST_SCHEMA, True))
+        table_store=TableStoreDB(dbconn, 'tbl_bad_data', TEST_SCHEMA, True)
+    )
 
-    tbl.store(TEST_DF)
+    tbl.store_chunk(TEST_DF)
 
     def inc_func(df):
         df_good = df[df['id'].isin(good_ids)]
         df_bad = df[df['id'].isin(bad_ids)]
         return df_good, df_bad
 
-    inc_process_many(ms, [tbl], [tbl_good, tbl_bad], inc_func)
+    inc_process_many(ds, [tbl], [tbl_good, tbl_bad], inc_func)
     assert_datatable_equal(tbl, TEST_DF)
     assert_datatable_equal(tbl_good, TEST_DF.loc[good_ids])
     assert_datatable_equal(tbl_bad, TEST_DF.loc[bad_ids])
 
     # Check this not delete the tables
-    inc_process_many(ms, [tbl], [tbl_good, tbl_bad], inc_func)
+    inc_process_many(ds, [tbl], [tbl_good, tbl_bad], inc_func)
     assert_datatable_equal(tbl, TEST_DF)
     assert_datatable_equal(tbl_good, TEST_DF.loc[good_ids])
     assert_datatable_equal(tbl_bad, TEST_DF.loc[bad_ids])
@@ -589,17 +584,15 @@ def test_error_handling(dbconn) -> None:
     GOOD_IDXS1 = [0, 1, 2, 3, 4, 5]
     CHUNKSIZE = 2
 
-    ms = MetaStore(dbconn)
+    ds = DataStore(dbconn)
 
-    tbl = DataTable(
+    tbl = ds.create_table(
         'tbl',
-        meta_table=ms.create_meta_table('tbl'),
         table_store=TableStoreDB(dbconn, 'tbl1_data', TEST_SCHEMA, True)
     )
 
-    tbl_good = DataTable(
+    tbl_good = ds.create_table(
         'tbl_good',
-        meta_table=ms.create_meta_table('tbl_good'),
         table_store=TableStoreDB(dbconn, 'tbl_good_data', TEST_SCHEMA, True)
     )
 
@@ -638,7 +631,7 @@ def test_error_handling(dbconn) -> None:
         return df
 
     inc_process_many(
-        ms,
+        ds,
         [tbl],
         [tbl_good],
         inc_func_bad,
@@ -648,7 +641,7 @@ def test_error_handling(dbconn) -> None:
     assert_datatable_equal(tbl_good, TEST_DF.loc[[0, 1, 2, 4, 5]])
 
     inc_process_many(
-        ms,
+        ds,
         [tbl],
         [tbl_good],
         inc_func_good,
@@ -666,7 +659,7 @@ def test_error_handling(dbconn) -> None:
     assert_datatable_equal(tbl, TEST_DF.loc[GOOD_IDXS1])
 
     inc_process_many(
-        ms,
+        ds,
         [tbl],
         [tbl_good],
         inc_func_bad,